Add Initial Rust Server Work
Adds a simple echo tcp server written in rust. These servers will be eventually extended into core components of the architecture.
The device-manager project is responsible for handling communication between modalities on a single device and the central ai server.
The server project handles communications with all registered devices and the various ai components and applications that are added onto it