Add Ability to Specify Config Args in Plugin Scripts
Adds the ability to specify a command line interface for accepting config options within custom plugins. The loader utilizes a 'cmd_args.yaml' file to create the command parser and passes the resulting dict to the plugin constructor.
Adds the ability for the loader to accept command line args without interfering with plugin args