Plugin Event Rewrite Exploration
We need to rewrite the plugin system to enable waiting on message responses in order to utilize their results. This is an initial exploration of a plugin system to add in the async-event system necessary for this addition.
We're also changing the process for handling messages in custom plugins from manually dispatching to having an automated registration system.