Investigate tokio-codec
One warning when compiling the server dependency is that the `length_delimited::*` stuff is deprecated in favor of `tokio-codec`. Unfortunately, we cannot switch over to this library because our `WriteJson` dependency requires the stream to "be" a `BytesMut`, while the tokio-codec produces a `Bytes` stream (this is inherited from the tokio_serde project).