Change Cli App to Have Command Subsystem
Modifies the cli app to be able to operate as a command line interface if desired. Currently, correct operation requires an internal knowledge of how the handle in question orders and treats it's arguments, as the parsed command line is sent "as-is". We have not yet decided on a system to allow for generic apps to communicate/register argparsers to allow for more "elaborate" constructions.

Recognized commands are in the form of "(<role>:)?<handle> <args...>". Additionally, cli operation commands of ":<command> <args...>" are also accepted.