Allow Endpoint Registration to Fail
Modifies endpoint registration code to allow for the registry to report registration failure. This currently only happens if a service tries to register an endpoint until a handle that is already registered (NOTE: May not be an error condition in the future).
Removes `Dispatcher::add_service` registration as it is the new workflow does not utilize it and it is somewhat unnecessary to have
Removes `Service::endpoints` function as it is no longer used
Changes signature of `Service::register_endpoints` to return an Error if a registration fails
Changes `rpc_schema!` to also derive Debug