Add Peer Identification in Rpc Callbacks
Adds a `caller` argument to all rpc endpoints. This value is assigned the socket address of the client that invoked the rpc. This allows us to access information about the client, particularly any read/write channels if we want to setup service callbacks.
Removes the `app_id` field from the Message type as it won't be used for identification