Improve Code Quality
Refactors the recent Client code to be more in line with current style decisions
Refactors the error handling in client exit callbacks to exist. Callbacks should now return a Result<(), Error> and errors will be propagated to the caller. An error in one callback will not stop the running of other callbacks.
Adds macro syntax to recognize attributes on rpc endpoints (currently unused)