Add Message Forwarding
Adds the ability for the read workflow to recognize that a particular message has been handled and just needs to be sent to the original client.
NOTE: This functionality currently does not work because of the usage of `try_recv` in the handler closure. This is currently necessary because of the typing of dispatch handlers due to the current synchronous implementation.