Wrap Serialization Code Into Classes
Wraps the code responsible for translating rpc messages between the transmission layer and the "rpc message" layer into two classes: rpc.BaseMessage and protocol.Protocol. rpc.BaseMessage is responsible for representing the rpc "message" as a logical structure and for defining how to serialize to/from a dictionary. protocol.Protocol defines how to take an rpc.BaseMessage and transform it into bytes for transmission on the socket
Adds in python type annotation support