Add Communication Handling
Adds reader/writer thread functions. These functions will be spawned on separate threads and automatically handle and route received/sent messages appropriately.
Adds a CommunicationHandler class to manage the interactions between the plugin and the reader/writer threads. This is to allow for the usage of asynchronous server requests by the plugin. This differs from the ConnectionHandler which deals with the specifics of the network communication from the reader/writer threads