Add Plugin Type Heirarchy
Adds in the concept of a Plugin type heirarchy and moves the old client code to residing in Client.main. Plugin.main will hold any code that needs to be run semi-regularly, such as the client implementation for Clients. For AppServers, it allows any persistant code to be run (such as re-indexing, etc.) - although the default implementation is good enough.