Integrate Registration and Dispatch into AppServer.main
Moves service endpoint registration (in the server) into an automatic call of AppServer.main (performed in `_register`).  This will require that all services for a modality to successfully register in order for the modality to run
Removes the usage of a dispatch map to allow dispatch to only consider endpoints that actually got registered.  Moves dispatch code into a separate module