Fix AppService Responses Being Nested When Returned
Fixes app service forwarding callback mistakenly treating the entire message as a response instead of just the response. The dispatcher code assumes the value returned from the dispatch handle is the response and takes that json and puts it in `msg.resp`. The app service callbacks were returning the whole message, instead of just the response field.
Fixes code so that mypy (mostly) type checks - it still complains about PluginBase not being a "name"